Don’t polyfill hashchange
I found that the ba-hashchange plugin was needlessly using a polling
mechanism in Safari (which supports the hashchange event in any case),
and that all recent browsers support it natively, so this commit removes
the polyfill and simply uses the native event. Any browser that does not
have such support, as reported by Modernizr, is redirected to fail.html.