Tight borders
An implementation of the idea in issue #33.

This is not yet a net improvement on the current live version, because
countries that become a LOT larger in certain views (GB in Historical,
for example) look oversimplified, and the manual path finagling we
are currently doing helps with that issue as well as with the border
infidelities.

However, it’s obviously possible to manually compensate for that, if we
adjust the simplification threshold based on the degree of maximal expansion
of a particular path; and I think when that is done the result will be
better than what we currently have. It will also simplify future cartogram
projects, by removing the need for painstaking manual tweaking of the
path simplification parameters.