Use Raphael rather than plain SVG
This is an experiment with using Raphaël as an abstraction layer
over SVG. The advantage is that it would work in IE, because Raphaël
falls back to VML if SVG is not supported.

The disadvantage – which is a pretty serious one – is that the
animations are unconscionably slow and jerky. If we can find a
solution to that problem, it could be a goer.