Make it runnable in IE 8
Note that some of these changes render it *in*compatible with other
browsers. In particular the current version of jQuery will not do
the right thing for selectors like $(".country") where the elements
with the "country" class are SVG paths. That’s the reason we used
getElementsByClassName in the first place, of course, but since IE 8
doesn’t support that method I changed it to jQuery for testing purposes.

The layout is pretty screwed in IE 8, though of course that is fixable.
The harder problem is that the animations are painfully, unreasonably
slow, even more so than in a modern browser, which I think makes this
approach a failed experiment.