Start animations in a tighter loop
Starting the animations in a tighter loop, i.e. reducing the delay
between starting the animation for different countries, does indeed
seem to subtly improve the appearance of the transitions (at least
in Safari). The effect whereby some borders appear to change in
thickness during transitions is completely eliminated by this
change, on my computer.

Closes #43.