Ignore *.cart.* files
I have a bunch of files matching this pattern in my local repo,
which were generated by an instrumented version of the Gastner-Newman
software to visualise the progress of the diffusion process.

I don’t want to check them into git, but this is a convenient location
to keep them, so ignoring them seems sensible.